Parse.h separation+ methods
