+++ structures/Aggregatestats.h
@@ -6,7 +6,7 @@
 #include "../imports/json.hpp"
 #include <fstream>
 #include <iostream>
-
+#include "../Utils.cpp"
 #ifndef TETR_IO_VISULIZER_AGGREGATESTATS_H
 #define TETR_IO_VISULIZER_AGGREGATESTATS_H
 
@@ -15,6 +15,15 @@ public:
     double apm;
     double pps;
     double vsscore;
+
+    Aggregatestats getAggregatestats(nlohmann::json const &json) {
+        auto aggregatestats = Aggregatestats(
+                to_double(json["apm"]),
+                to_double(json["ppm"]),
+                to_double(json["vsscore"])
+        );
+        return aggregatestats;
+    }
 };
