+++ structures/Clears.h
@@ -6,6 +6,8 @@
 #include "../imports/json.hpp"
 #include <fstream>
 #include <iostream>
+#include "../Utils.cpp"
+
 #ifndef TETR_IO_VISULIZER_CLEARS_H
 #define TETR_IO_VISULIZER_CLEARS_H
 
@@ -27,8 +29,28 @@ public:
     int tspinquads;
     int tspinpentas;
     int allclear;
-};
 
+    Clears getClears(nlohmann::json const &json) {
+        auto clears = Clears(
+                to_int(json["singles"]),
+                to_int(json["doubles"]),
+                to_int(json["triples"]),
+                to_int(json["quads"]),
+                to_int(json["pentas"]),
+                to_int(json["realtspins"]),
+                to_int(json["minitspins"]),
+                to_int(json["minitpinsingles"]),
+                to_int(json["tspinsingles"]),
+                to_int(json["minitspindoubles"]),
+                to_int(json["tspindoubles"]),
+                to_int(json["tspintriples"]),
+                to_int(json["tspinquads"]),
+                to_int(json["tspinpentas"]),
+                to_int(json["allclear"])
+        );
+        return clears;
+    }
+};
 
 
 #endif //TETR_IO_VISULIZER_CLEARS_H
