+++ structures/ConstantsOverrides.h
@@ -14,6 +14,10 @@
 class Constants_overrides {
 public:
 
+    Constants_overrides getConstants_overrides(nlohmann::json const &json) {
+        auto constants_overrides = Constants_overrides();
+        return constants_overrides;
+    }
 };
