+++ structures/Controlling.h
@@ -5,11 +5,13 @@
 #include <vector>
 #include "../imports/json.hpp"
 #include <fstream>
+#include "../Utils.cpp"
 #include <iostream>
+
 #ifndef TETR_IO_VISULIZER_CONTROLLING_H
 #define TETR_IO_VISULIZER_CONTROLLING_H
 
-class Controlling{
+class Controlling {
 public:
     int ldas;
     int ldasiter;
@@ -19,6 +21,21 @@ public:
     bool rshift;
     int lastshift;
     bool softdrop;
+
+    Controlling getControlling(nlohmann::json const &json) {
+        auto controlling = Controlling(
+                to_int(json["ldas"]),
+                to_int(json["ldasiter"]),
+                to_bool(json["lshift"]),
+                to_int(json["rdas"]),
+                to_int(json["rdasiter"]),
+                to_bool(json["rshift"]),
+                to_int(json["lastshift"]),
+                to_bool(json["softdrop"])
+        );
+
+        return controlling;
+    }
 };
