+++ structures/Finesse.h
@@ -6,14 +6,25 @@
 #include "../imports/json.hpp"
 #include <fstream>
 #include <iostream>
+#include "../Utils.cpp"
+
 #ifndef TETR_IO_VISULIZER_FINESSE_H
 #define TETR_IO_VISULIZER_FINESSE_H
 
-class Finesse{
+class Finesse {
 public:
     int combo;
     int faults;
     int perfectpieces;
+
+    Finesse getFinesse(nlohmann::json const &json) {
+        auto finesse = Finesse(
+                to_int(json["combo"]),
+                to_int(json["faults"]),
+                to_int(json["perfectpieces"])
+        );
+        return finesse;
+    }
 };
