+++ structures/Handling.h
@@ -6,6 +6,7 @@
 #include "../imports/json.hpp"
 #include <fstream>
 #include <iostream>
+#include "../Utils.cpp"
 #ifndef TETR_IO_VISULIZER_HANDLING_H
 #define TETR_IO_VISULIZER_HANDLING_H
 
@@ -19,6 +20,19 @@ public:
     bool safelock;
     bool cancel;
     bool may20g;
+
+    Handling getHandling(nlohmann::json const &json) {
+        auto handling = Handling(
+                to_int(json["arr"]),
+                to_int(json["das"]),
+                to_int(json["dcd"]),
+                to_int(json["sdf"]),
+                to_bool(json["safelock"]),
+                to_bool(json["cancel"]),
+                to_bool(json["may20g"])
+        );
+        return handling;
+    }
 };
