+++ structures/Hold.h
@@ -6,16 +6,25 @@
 #include "../imports/json.hpp"
 #include <fstream>
 #include <iostream>
+#include "../Utils.cpp"
+
 #ifndef TETR_IO_VISULIZER_HOLD_H
 #define TETR_IO_VISULIZER_HOLD_H
 
 
-class Hold{
+class Hold {
 public:
     std::string piece;
     bool locked;
-};
 
+    Hold getHold(nlohmann::json const &json) {
+        auto hold = Hold(
+                to_string(json["piece"]),
+                to_bool(json["locked"])
+        );
+        return hold;
+    }
+};
 
 
 #endif //TETR_IO_VISULIZER_HOLD_H
