+++ structures/IgeData.h
@@ -7,10 +7,13 @@
 #include "Target.h"
 #include <fstream>
 #include <iostream>
+#include "IntData.h"
+#include "../Utils.cpp"
+
 #ifndef TETR_IO_VISULIZER_IGEDATA_H
 #define TETR_IO_VISULIZER_IGEDATA_H
 
-class IgeData{
+class IgeData {
 public:
     std::string type;
     std::string gameid;
@@ -21,7 +24,22 @@ public:
     bool hoisted;
     int subframe;
 
-    IgeData data;
+    IntData data;
+
+    IgeData getIgeData(nlohmann::json const &json) {
+        auto igedata = IgeData(
+                to_string(json["type"]),
+                to_string(json["gameid"]),
+                getTarget(json["targets"]),
+                to_int(json["frame"]),
+                to_bool(json["value"]),
+                to_string(json["key"]),
+                to_bool(json["hoisted"]),
+                to_int(json["subframe"]),
+                getIntData(json["data"])
+        );
+        return igedata;
+    }
 };
 
 #endif //TETR_IO_VISULIZER_IGEDATA_H
