+++ structures/IntData.h
@@ -6,6 +6,7 @@
 #include "../imports/json.hpp"
 #include <fstream>
 #include <iostream>
+#include "../Utils.cpp"
 #ifndef TETR_IO_VISULIZER_INTDATA_H
 #define TETR_IO_VISULIZER_INTDATA_H
 
@@ -20,6 +21,20 @@ public:
     int y;
     int size;
     std::string username;
+
+    IntData getIntData(nlohmann::json const &json) {
+        auto intData = IntData(
+                to_int(json["iid"]),
+                to_string(json["type"]),
+                to_int(json["amt"]),
+                to_int(json["ackiid"]),
+                to_int(json["x"]),
+                to_int(json["y"]),
+                to_int(json["size"]),
+                to_string(json["username"])
+        );
+        return intData;
+    }
 };
 
 #endif //TETR_IO_VISULIZER_INTDATA_H
