+++ structures/Objective.h
@@ -6,12 +6,18 @@
 #include "../imports/json.hpp"
 #include <fstream>
 #include <iostream>
+
 #ifndef TETR_IO_VISULIZER_OBJECTIVE_H
 #define TETR_IO_VISULIZER_OBJECTIVE_H
 
 class Objective {
 public:
     std::string type;
+
+    Objective getObjective(nlohmann::json const &json) {
+        auto objective = Objective(to_string(json["type"]));
+        return objective;
+    }
 };
 
 #endif //TETR_IO_VISULIZER_OBJECTIVE_H
