+++ structures/Player.h
@@ -10,6 +10,8 @@
 #include "ExtraAvgTracking.h"
 #include <fstream>
 #include <iostream>
+#include "../Utils.cpp"
+
 #ifndef TETR_IO_VISULIZER_PLAYER_H
 #define TETR_IO_VISULIZER_PLAYER_H
 
@@ -30,6 +32,35 @@ public:
     AvgTracking secondaryAvgTracking;
     AvgTracking tertiaryAvgTracking;
     ExtraAvgTracking extraAvgTracking;
+
+    std::vector<Player> getEndcontext(nlohmann::json const &fulljson) {
+        auto endcontext = std::vector<Player>();
+        nlohmann::json json;
+        int i = 0;
+
+        while (i < fulljson.size()) {
+            json = fulljson[i];
+            endcontext.push_back(
+                    Player(to_string(json["id"]),
+                           to_string(json["username"]),
+                           Handling().getHandling(json["handling"]),
+                           to_bool(json["active"]),
+                           to_bool(json["success"]),
+                           to_int(json["inputs"]),
+                           to_int(json["pieceplaced"]),
+                           to_int(json["naturalorder"]),
+                           to_int(json["score"]),
+                           to_int(json["wins"]),
+                           Points().getPoints(json["points"]),
+                           getSecondaryAvgTracking(json["secondaryAvgTracking"]),
+                           getTertiaryAvgTracking(json["tertiaryAvgTracking"]),
+                           getExtraAvgTracking(json["extraAvgTracking"])));
+
+            i++;
+        }
+
+        return endcontext;
+    }
 };
