+++ structures/Points.h
@@ -7,6 +7,7 @@
 #include "../imports/json.hpp"
 #include <fstream>
 #include <iostream>
+#include "../Utils.cpp"
 #ifndef TETR_IO_VISULIZER_POINTS_H
 #define TETR_IO_VISULIZER_POINTS_H
 
@@ -16,6 +17,16 @@ public:
     double secondary;
     double tertiary;
     PointsExtra extra;
+
+    Points getPoints(nlohmann::json const &json) {
+        auto points = Points(
+                to_int(json["primary"]),
+                to_double(json["secondary"]),
+                to_double(json["tertiary"]),
+                PointsExtra().getPointsExtra(json['extra'])
+        );
+        return points;
+    }
 };
