+++ structures/PointsExtra.h
@@ -6,12 +6,18 @@
 #include "../imports/json.hpp"
 #include <fstream>
 #include <iostream>
+
 #ifndef TETR_IO_VISULIZER_POINTSEXTRA_H
 #define TETR_IO_VISULIZER_POINTSEXTRA_H
 
 class PointsExtra {
 public:
     std::string vs;
+
+    PointsExtra getPointsExtra(nlohmann::json const &json) {
+        auto pointsExtra = PointsExtra(to_string(json["vs"]));
+        return pointsExtra;
+    }
 };
