+++ structures/Stats.h
@@ -37,6 +37,33 @@ public:
     Garbage garbage;
     int kills;
     Finesse finesse;
+
+    Stats getStats(nlohmann::json const &json) {
+        auto stats = Stats(
+                to_long(json["seed"]),
+                to_int(json["lines"]),
+                to_int(json["level_lines"]),
+                to_int(json["level_lines_needed"]),
+                to_int(json["inputs"]),
+                to_int(json["holds"]),
+                Time().getTime(json["time"]),
+                to_int(json["score"]),
+                to_int(json["zenlevel"]),
+                to_int(json["zenprogress"]),
+                to_int(json["level"]),
+                to_int(json["combo"]),
+                to_int(json["btb"]),
+                to_int(json["topbtb"]),
+                to_int(json["currentbtbchainpower"]),
+                to_int(json["tspins"]),
+                to_int(json["piecesplaced"]),
+                Clears().getClears(json["clears"]),
+                Garbage().getGarbage(json["garbage"]),
+                to_int(json["kills"]),
+                Finesse().getFinesse(json["finesse"])
+        );
+        return stats;
+    }
 };
