+++ structures/Target.h
@@ -6,12 +6,24 @@
 #include "../imports/json.hpp"
 #include <fstream>
 #include <iostream>
+
 #ifndef TETR_IO_VISULIZER_TARGET_H
 #define TETR_IO_VISULIZER_TARGET_H
 
-class Target{
+class Target {
 public:
 
+    std::vector<Target> getTarget(nlohmann::json const &fulljson) {
+        auto target = std::vector<Target>();
+        nlohmann::json json;
+        int i = 0;
+        while (i < fulljson.size()) {
+            json = fulljson[i];
+            target.push_back(Target());
+            i++;
+        }
+        return target;
+    }
 };
 
 #endif //TETR_IO_VISULIZER_TARGET_H
