+++ structures/Time.h
@@ -4,6 +4,7 @@
 #include <string>
 #include <vector>
 #include "../imports/json.hpp"
+#include "../Utils.cpp"
 #include <fstream>
 #include <iostream>
 #ifndef TETR_IO_VISULIZER_TIME_H
@@ -16,6 +17,17 @@ public:
     bool locked;
     int prev;
     int framoffset;
+
+    Time getTime(nlohmann::json const &json) {
+        auto time = Time(
+                to_int(json["start"]),
+                to_bool(json["zero"]),
+                to_bool(json["locked"]),
+                to_int(json["prev"]),
+                to_int(json["framoffset"])
+        );
+        return time;
+    }
 };
